# Backend/docs/swagger.yaml
info:
  contact: {}
paths:
  /notification/payment/cancelled:
    post:
      responses: {}
  /notification/payment/created:
    post:
      responses: {}
  /notification/payment/successful:
    post:
      responses: {}
  /payment:
    get:
      parameters:
      - description: Order UUID
        in: query
        name: order_uuid
        type: string
      - description: Payment UUID
        in: query
        name: payment_uuid
        type: string
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
  /payment/callback/stripe/{payment_uuid}:
    get:
      parameters:
      - description: Payment UUID
        in: path
        name: payment_uuid
        required: true
        type: string
      - description: Checkout Session ID
        in: path
        name: checkout_session_id
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /payment/init:
    post:
      parameters:
      - description: Amount
        in: formData
        name: amount
        required: true
        type: integer
      - description: Order UUID
        in: formData
        name: order_uuid
        required: true
        type: string
      - description: Method
        in: formData
        name: method
        required: true
        type: string
      - description: Currency
        in: formData
        name: currency
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /payment/update:
    patch:
      parameters:
      - description: Payment UUID
        in: formData
        name: payment_uuid
        required: true
        type: string
      - description: Status
        in: formData
        name: status
        required: true
        type: string
      - description: Reference ID
        in: formData
        name: reference_id
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /user:
    get:
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
  /user/email/verification:
    post:
      parameters:
      - description: Email Verification Token
        in: formData
        name: email_verify_token
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /user/login:
    post:
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /user/logout:
    post:
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
  /user/register:
    post:
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
  /user/token:
    post:
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
swagger: "2.0"